pipe and command execution files are semi norminated .. fixed an issue with pipe command execution
